% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellorustc-package.R
\docType{package}
\name{hellorustc-package}
\alias{hellorustc}
\alias{hellorustc-package}
\title{hellorustc: Use `rustc` only with R}
\description{
Proof of concept for using `rustc` directly with R packages.
}
\author{
\strong{Maintainer}: Mossa Reimert \email{mossa@sund.ku.dk} (\href{https://orcid.org/0009-0007-9297-1523}{ORCID})

}
\keyword{internal}
