% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellorustc-package.R
\docType{package}
\name{hellorustc-package}
\alias{hellorustc}
\alias{hellorustc-package}
\title{hellorustc: Proof of concept for using \code{rustc} directly with R packages}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mossa Reimert \email{mossa@sund.ku.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
