% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellorustc-package.R
\docType{package}
\name{hellorustc-package}
\alias{hellorustc}
\alias{hellorustc-package}
\title{hellorustc: Use `rustc` only with R}
\description{
Proof of concept for using \code{rustc} directly with R packages
}
\author{
\strong{Maintainer}: Mossa Reimert \email{mossa@sund.ku.dk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
